OSGi support - we now export OSGi metadata in our manifest.
The tool we're currently using is somewhat crufty. It attempts to parse the .java files to figure out what needs importing and exporting. As a consequence, when we import inner classes, it breaks. For example, it will do the wrong thing if it sees this statement:
  import com.google.inject.InjectorImpl.SingleParameterInjector;
The alternative is a popular tool called BND. It's what most of the OSGi diehards use, and we might consider switching to it. But it is significantly more complex (it has its own property language) and requires .bnd files for each exported .jar. Regardless, that tool is smarter and if this strategy doesn't work out, we should adopt that tool.

git-svn-id: http://google-guice.googlecode.com/svn/trunk@501 d779f126-a31b-0410-b53b-1d3aecad763e