Stuart's fixes for the BytecodeGenTest.
Also changing Classes.isInnerClass to do the modifiers check first. Turns out that by calling Class.getEnclosingClass, we were causing the enclosing class to get loaded. Not really a big deal, but not something we really want to do.

Still to do -- change the code to not use this classloader if any intercepted method is package-private.

git-svn-id: http://google-guice.googlecode.com/svn/trunk@542 d779f126-a31b-0410-b53b-1d3aecad763e