Removing the obsolete commands extension. Almost all of it has been rolled into the Guice SPI.
The only remining file (InterceptingInjectorBuilder) now exists for maintenance on a new project that I set up for my personal use:
http://code.google.com/p/publicobject/
I'll have to copy this into Guiceberry which is currently the only project I know of that depends on this file.

My biggest problem with IIB is that it's a leaky abstraction. Unfortunately many bindings cannot be intercepted - anything that binds directly to a constructor is not eligible for interception. The approach taken by the guiceyfruit project (and potentially useful for inclusion in Guice proper) might be preferred. http://code.google.com/p/guiceyfruit/

git-svn-id: http://google-guice.googlecode.com/svn/trunk@748 d779f126-a31b-0410-b53b-1d3aecad763e