Removing all ASM and cglib dependencies from the no_aop build of Guice.
We now remove those .jar files from the build package, and remove references to them from our build in common.xml.

Some tests now fail because they're written to expect a specific class name in the exception message, like:
   "at com.google.inject.Foo.<init>(Foo.java:223)"
but without LineNumbers we show a message like:  
   "at com.google.inject.Foo.<init>(Unknown Source)"
These failures are annoying but not really a dealbreaker for me. I'd prefer to have failing tests in the no-aop build than to not have test coverage for line number inclusion.

git-svn-id: http://google-guice.googlecode.com/svn/trunk@852 d779f126-a31b-0410-b53b-1d3aecad763e